formatted source code

+++ contracts/Migrations.sol
@@ -1,12 +1,10 @@
 pragma solidity ^0.5.0;
 
-
 contract Migrations {
     address public owner;
-    uint public last_completed_migration; // solhint-disable-line
+    uint256 public last_completed_migration; // solhint-disable-line
 
-    constructor()
-    public {
+    constructor() public {
         owner = msg.sender;
     }
 
@@ -14,15 +12,12 @@ contract Migrations {
         if (msg.sender == owner) _;
     }
 
-    function setCompleted(uint completed)
-    public
-    restricted {
+    function setCompleted(uint256 completed) public restricted {
         last_completed_migration = completed;
     }
 
-    function upgrade(address new_address)
-    public
-    restricted { // solhint-disable-line
+    function upgrade(address new_address) public restricted {
+        // solhint-disable-line
         Migrations upgraded = Migrations(new_address);
         upgraded.setCompleted(last_completed_migration);
     }
